Added thunks for spotifyGetManyPlaylists
