+++ client/src/components/Oauth/spotifyApiReducer.js
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { spotifyProfileThunk } from './Spotify/spotifyApiThunks';
+import { spotifyProfileThunk, spotifyGetManyPlaylistsThunk } from './Spotify/spotifyApiThunks';
 
 // =================================================================================>
 // parameters for spotify OAuth
@@ -42,6 +42,7 @@ const spotifyApiSlice = createSlice({
         builder.addCase(spotifyProfileThunk.fulfilled, (state, action) => {
             state.profile = action.payload;
         });
+
     }
 });
