+++ client/src/components/home/redux/actionTypes.js
@@ -7,7 +7,9 @@ export const actionTypes = {
     ADD_SONG:'addSong',
     REMOVE_SONG:'removeSong',
     GET_SONGS:'getSongs',
-    TOGGLE_FAVE:'toggleFavoriteSong'
+    TOGGLE_FAVE:'toggleFavoriteSong',
+
+    SPOTIFY_GET_PLAYLISTS:'spotifyGetPlaylists',
 }
