+++ client/src/components/home/redux/playlistsReducer.js
@@ -1,13 +1,15 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { createPlaylistAsync,
+import {
+  createPlaylistAsync,
   deletePlaylistAsync,
   editPlaylistAsync,
   addSongAsync,
   getPlaylistsAsync,
   removeSongAsync,
   getSongsAsync,
-  getOnePlaylist
- } from './thunks';
+  getOnePlaylist,
+  spotifyGetManyPlaylistsThunk
+} from './thunks';
 
 /*
 ACTIONS:
@@ -74,6 +76,11 @@ const playlistsSlice = createSlice({
         playlist.songs = songs;
       }
     });
+
+
+    builder.addCase(spotifyGetManyPlaylistsThunk.fulfilled, (state, action) => {
+      console.log(action.payload);
+  });
   }
 });
