+++ client/src/components/home/redux/thunks.js
@@ -72,3 +72,15 @@ export const getSongsAsync = createAsyncThunk(
         return await PlaylistsService.getSongs(playlistID);
     }
 );
+
+
+
+
+export const spotifyGetManyPlaylistsThunk = createAsyncThunk(
+    actionTypes.SPOTIFY_GET_PLAYLISTS,
+    async ({playlists, accessToken}) => {
+      return await PlaylistsService.spotifyGetManyPlaylists(playlists, accessToken);
+    }
+);
+  
+
