+++ server/routes/spotify.js
@@ -1,5 +1,3 @@
-import { TYPE_ALBUM, TYPE_PLAYLIST, TYPE_TRACK, TYPE_SPOTIFY, TYPE_YOUTUBE } from '../shared/playlistTypeConstants';
-
 const express = require('express');
 var router = express.Router();
 
@@ -138,96 +136,5 @@ router.get('/profile', function (req, res) {
 
 });
 
-// fetches all tracks for a single playlist
-
-function getTracksHelper(access_token, next, playlist) {
-    // if next link is null, dont do anything
-    return next ? fetch(next, {
-        method: "GET",
-        headers: { Authorization: `Bearer ${access_token}` }
-    }).then(response => {
-        if (response.status === 200) {
-            return response.json();
-        } else {
-            return Promise.reject(response);
-        }
-    }).then(data => {
-        // console.log(data);
-        for (const i of data.items) {
-            playlist.songs.push(
-                {
-                    songID: uuid(),
-                    artist: i.track.artists[0].name,
-                    name: i.track.name,
-                    type: 'spotify',
-                    link: i.track.uri,
-                    imageLink: i.track.album.images[0].url,
-                    album: i.track.album.name,
-                    duration: i.track.duration_ms,
-                    releaseDate: i.track.album.release_date,
-                }
-            );
-        }
-        if (data.next) {
-            return getTracksHelper(access_token, data.next, playlist);
-        } else {
-            return playlist;
-        }
-    }) // don't catch, let error bubble up to route handler
-        :
-        playlist;
-}
-
-// TODO: test to make sure this doesn't get rate-limited on reasonably sized playlists
-router.post('/importManySpotify', async (req, res, next) => {
-    const { playlists, access_token } = req.body;
-
-    let queryUrl;
-
-    Promise.allSettled(playlists?.map(playlist => {
-
-        const type = playlist.type;
-
-        if   (type === TYPE_PLAYLIST) queryUrl = `https://api.spotify.com/v1/playlists/${playlist.id}`;
-        else if (type === TYPE_ALBUM) queryUrl = `https://api.spotify.com/v1/albums/${playlist.id}`;
-        else {
-            return res.status(400).send({ error: "one of the playlists ID's had invalid type (not playlist or album)" });
-        }
-
-        return fetch(queryUrl, {
-            method: "GET",
-            headers: { Authorization: `Bearer ${access_token}` }
-        }).then(response => {
-            if (response.status === 200) {
-                return response.json();
-            } else {
-                return Promise.reject(response);
-            }
-        }).then(data => getTracksHelper(access_token, data.tracks.href, {
-            playlistID: uuid(),
-            dateCreated: new Date(),
-            description: data.description,
-            name: data.name,
-            author: new ObjectId(), // TODO: objectid of the user who is importing the playlist
-            isFavorited: false,
-            coverImageURL: data.images[0].url,
-            songs: [],
-            originSpotifyId: data.id,
-            isAlbum: false,
-        })
-        ).then(async (playlist) => {
-            const result = await playlistsCol.insertOne(playlist);
-            console.log(`inserted ${result.insertedId}`);
-            return playlist.originSpotifyId; // useful for frontend retry
-        })
-    }))
-        .then(outcomes => {
-            if (!outcomes.some((o) => o.status === "fulfilled")) {
-                return res.status(500).send(outcomes);
-            }
-            return res.status(200).send(outcomes);
-        })
-    // NOTE: no catch, Promise.allSettled never rejects.
-});
 
 module.exports = router;
